rename: components아래 직접 컴포넌트를 포함하는 폴더들은 PascalCase로 변경
컴포넌트를 포함하는 폴더는 컴포넌트와 마찬가지로 PascalCase를 사용하는 것이
컴포넌트가 들어있다는 것을 바로 파악할 수 있으므로 변경하였습니다
