+++ src/HabitTracker.tsx
@@ -1,6 +1,6 @@
 import Header from 'components/Header/Header';
-import Controller from 'components/controller/Controller';
-import DashBoard from 'components/dashBoard/Dashboard';
+import Controller from 'components/Controller/Controller';
+import DashBoard from 'components/Dashboard/Dashboard';
 import TempHabitCreateForm from 'components/temporary/TempHabitCreateForm';
 import TempHabitList from 'components/temporary/TempHabitList';
 import styles from './HabitTracker.module.css';
